+++ lua/crates/command.lua
@@ -22,6 +22,7 @@ local sub_commands = {
 
     { "expand_plain_crate_to_inline_table", actions.expand_plain_crate_to_inline_table },
     { "extract_crate_into_table",           actions.extract_crate_into_table },
+    { "goto_workspace_definition",          require("crates.workspace").goto_definition },
 
     { "open_homepage",                      actions.open_homepage },
     { "open_repository",                    actions.open_repository },